requirements.txt
fastapi
uvicorn
requests
sendgrid
pillow
opencv-python
easyocr
numpy
imutils